added pic display method and checkbox process file
As a note i've been typing in database names and such manually so i
wouldn't suggest cutting and pasting that stuff. Also the
checkbox_test.php file does not add anything to the databases, it simply
displays the index of the checked off book enteries but this should be
able to be used to find books by index in the book database and insert
them into he wanted books section.